third_party/blink/web_tests/http/tests/images/feature-policy-unoptimized-images-expected.txt

CONSOLE ERROR: Feature policy violation: unoptimized-images is not allowed in this document.
